pyproject.toml
[build-system]
requires = ["flit_core ~=3.12"]
build-backend = "flit_core.buildapi"

[project]
name = "musicli-sequencer"
authors = [{name = "Aaron Friesen"}]
readme = "README.md"
license = {file = "LICENSE.txt"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
dynamic = ["version", "description"]
dependencies = [
	"mido ~=1.3.3",
	"pyfluidsynth ~=1.3.4",
]

[project.urls]
Home = "https://frie.dev/musicli"
GitHub = "https://github.com/friedev/musicli"
Changelog = "https://github.com/friedev/musicli/blob/main/CHANGELOG.md"

[project.scripts]
musicli = "musicli_sequencer.main:main"
